+++ client/recv.go
@@ -1,10 +1,12 @@
 package client
 
 import (
+	"crypto/tls"
 	"fmt"
 	"net"
 	"os"
 	"path/filepath"
+	"sync"
 	"time"
 
 	fio "github.com/fatedier/fft/pkg/io"
@@ -26,10 +28,12 @@ func (svc *Service) recvFile(id string, filePath string) error {
 	if err != nil {
 		return err
 	}
+	conn = tls.Client(conn, &tls.Config{InsecureSkipVerify: true})
 	defer conn.Close()
 
 	msg.WriteMsg(conn, &msg.ReceiveFile{
-		ID: id,
+		ID:         id,
+		CacheCount: int64(svc.cacheCount),
 	})
 
 	conn.SetReadDeadline(time.Now().Add(10 * time.Second))
@@ -66,12 +70,15 @@ func (svc *Service) recvFile(id string, filePath string) error {
 	}
 	defer f.Close()
 
+	var wait sync.WaitGroup
 	count := m.Fsize
 	bar := pb.New(int(count))
 	bar.ShowSpeed = true
 	bar.SetUnits(pb.U_BYTES)
 
-	bar.Start()
+	if !svc.debugMode {
+		bar.Start()
+	}
 
 	callback := func(n int) {
 		bar.Add(n)
@@ -79,61 +86,85 @@ func (svc *Service) recvFile(id string, filePath string) error {
 
 	recv := receiver.NewReceiver(0, fio.NewCallbackWriter(f, callback))
 	for _, worker := range m.Workers {
-		addr := worker
-		go newRecvStream(recv, id, addr, svc.debugMode)
+		wait.Add(1)
+		go func(addr string) {
+			newRecvStream(recv, id, addr, svc.debugMode)
+			wait.Done()
+		}(worker)
 	}
-	recv.Run()
-	bar.Finish()
+
+	recvDoneCh := make(chan struct{})
+	streamCloseCh := make(chan struct{})
+	go func() {
+		recv.Run()
+		close(recvDoneCh)
+	}()
+	go func() {
+		wait.Wait()
+		close(streamCloseCh)
+	}()
+
+	select {
+	case <-recvDoneCh:
+	case <-streamCloseCh:
+		select {
+		case <-recvDoneCh:
+		case <-time.After(2 * time.Second):
+		}
+	}
+
+	if !svc.debugMode {
+		bar.Finish()
+	}
 	return nil
 }
 
 func newRecvStream(recv *receiver.Receiver, id string, addr string, debugMode bool) {
-	first := true
-	for {
-		if !first {
-			time.Sleep(3 * time.Second)
-		} else {
-			first = false
-		}
+	conn, err := net.Dial("tcp", addr)
+	if err != nil {
+		log(debugMode, "[%s] %v", addr, err)
+		return
+	}
+	conn = tls.Client(conn, &tls.Config{InsecureSkipVerify: true})
 
-		conn, err := net.Dial("tcp", addr)
+	msg.WriteMsg(conn, &msg.NewReceiveFileStream{
+		ID: id,
+	})
+
+	conn.SetReadDeadline(time.Now().Add(10 * time.Second))
+	raw, err := msg.ReadMsg(conn)
+	if err != nil {
+		conn.Close()
+		log(debugMode, "[%s] %v", addr, err)
+		return
+	}
+	conn.SetReadDeadline(time.Time{})
+	m, ok := raw.(*msg.NewReceiveFileStreamResp)
+	if !ok {
+		conn.Close()
+		log(debugMode, "[%s] read NewReceiveFileStreamResp format error", addr)
+		return
+	}
+
+	if m.Error != "" {
+		conn.Close()
+		log(debugMode, "[%s] new recv file stream error: %s", addr, m.Error)
+		return
+	}
+
+	s := stream.NewFrameStream(conn)
+	for {
+		frame, err := s.ReadFrame()
 		if err != nil {
-			log(debugMode, "[%s] %v", addr, err)
 			return
 		}
-
-		msg.WriteMsg(conn, &msg.NewReceiveFileStream{
-			ID: id,
+		recv.RecvFrame(frame)
+		err = s.WriteAck(&stream.Ack{
+			FileID:  frame.FileID,
+			FrameID: frame.FrameID,
 		})
-
-		conn.SetReadDeadline(time.Now().Add(10 * time.Second))
-		raw, err := msg.ReadMsg(conn)
 		if err != nil {
-			conn.Close()
-			log(debugMode, "[%s] %v", addr, err)
-			continue
-		}
-		conn.SetReadDeadline(time.Time{})
-		m, ok := raw.(*msg.NewReceiveFileStreamResp)
-		if !ok {
-			conn.Close()
-			log(debugMode, "[%s] read NewReceiveFileStreamResp format error", addr)
-			continue
-		}
-
-		if m.Error != "" {
-			conn.Close()
-			log(debugMode, "[%s] new recv file stream error: %s", addr, m.Error)
-			continue
-		}
-
-		s := stream.NewFrameStream(conn)
-		for {
-			frame, err := s.ReadFrame()
-			if err != nil {
-				return
-			}
-			recv.RecvFrame(frame)
+			return
 		}
 	}
 }