+++ worker/service.go
@@ -1,7 +1,13 @@
 package worker
 
 import (
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/tls"
+	"crypto/x509"
+	"encoding/pem"
 	"fmt"
+	"math/big"
 	"net"
 	"strconv"
 	"time"
@@ -31,8 +37,9 @@ type Service struct {
 	serverAddr     string
 	advicePublicIP string
 
-	l        net.Listener
-	matchCtl *MatchController
+	l         net.Listener
+	matchCtl  *MatchController
+	tlsConfig *tls.Config
 }
 
 func NewService(options Options) (*Service, error) {
@@ -56,8 +63,9 @@ func NewService(options Options) (*Service, error) {
 		serverAddr:     options.ServerAddr,
 		advicePublicIP: options.AdvicePublicIP,
 
-		l:        l,
-		matchCtl: NewMatchController(),
+		l:         l,
+		matchCtl:  NewMatchController(),
+		tlsConfig: generateTLSConfig(),
 	}, nil
 }
 
@@ -69,6 +77,7 @@ func (svc *Service) Run() error {
 	if err != nil {
 		return err
 	}
+	conn = tls.Client(conn, &tls.Config{InsecureSkipVerify: true})
 
 	_, portStr, err := net.SplitHostPort(svc.l.Addr().String())
 	if err != nil {
@@ -96,6 +105,7 @@ func (svc *Service) worker() error {
 		if err != nil {
 			return err
 		}
+		conn = tls.Server(conn, svc.tlsConfig)
 		go svc.handleConn(conn)
 	}
 }
@@ -141,4 +151,25 @@ func (svc *Service) handleConn(conn net.Conn) {
 		conn.Close()
 		return
 	}
+}
+
+// Setup a bare-bones TLS config for the server
+func generateTLSConfig() *tls.Config {
+	key, err := rsa.GenerateKey(rand.Reader, 1024)
+	if err != nil {
+		panic(err)
+	}
+	template := x509.Certificate{SerialNumber: big.NewInt(1)}
+	certDER, err := x509.CreateCertificate(rand.Reader, &template, &template, &key.PublicKey, key)
+	if err != nil {
+		panic(err)
+	}
+	keyPEM := pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(key)})
+	certPEM := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: certDER})
+
+	tlsCert, err := tls.X509KeyPair(certPEM, keyPEM)
+	if err != nil {
+		panic(err)
+	}
+	return &tls.Config{Certificates: []tls.Certificate{tlsCert}}
 }