+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ### v0.1.0
 
 - Support upload and download process bar.
+- Enable TLS between all components.
+- Support ACK message and retry missing data.
+- Add `-n` to set frame size of sender. Add `-c` to set cache frame count.
 
 ### v0.0.1